Feature: Selector de empleados

+++ ContaGestion-api/src/main/java/com/exactian/gestion/repositories/EmpleadoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.exactian.gestion.model.Empleado;
 
 
-public interface EmpleadoRepository  extends JpaRepository<Empleado, Integer>{
+public interface EmpleadoRepository  extends JpaRepository<Empleado, Long>{
 
 }
